+++ pytsite/assetman/__init__.py
@@ -4,7 +4,7 @@
 from . import _error as error
 from ._api import register_package, library, preload, remove, dump_js, dump_css, url, add_inline, dump_inline, \
     get_urls, get_locations, reset, detect_collection, build, is_package_registered, register_global, _add_task, \
-    t_browserify, t_copy, t_copy_static, t_less, t_js, t_css, js_module
+    t_browserify, t_copy, t_copy_static, t_less, t_js, t_css, js_module, setup
 
 __author__ = 'Alexander Shepetko'
 __email__ = 'a@shepetko.com'
@@ -13,7 +13,7 @@ __license__ = 'MIT'
 
 def _init():
     from os import path
-    from pytsite import reg, console, events, lang, tpl, router
+    from pytsite import reg, console, events, lang, tpl, router, setup as pytsite_setup
     from . import _console_command, _api
 
     # Registry variables
@@ -29,6 +29,7 @@ def _init():
     # Event handlers
     router.on_dispatch(reset, -999)
     router.on_xhr_dispatch(reset, -999, 'post')  # Workaround for forms
+    pytsite_setup.on_setup(setup)
     events.listen('pytsite.update.after', lambda: build(switch_maintenance=False))
 
     # Tpl resources
@@ -41,7 +42,7 @@ def _init():
     # Register assetman itself and add required assets for all pages
     register_package(__name__)
     js_module('assetman', __name__ + '@assetman')
-    t_js(__name__ + '@**/*.js')
+    t_js(__name__ + '@**')
     preload(__name__ + '@require.js', True)
     preload(__name__ + '@require-config.js', True)